pyproject.toml
[tool.poetry]
name = "xrp"
version = "0.1.0"
description = ""
authors = ["Larry O'Brien <lobrien@knowing.net>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
numpy = "^2.1.2"
opencv-python = "^4.10.0.84"
robotpy = "^2024.3.2.2"
robotpy-xrp = "^2024.3.2.1"
wpilib = "^2024.3.2.1"
pynput = "^1.7.7"
rubicon-objc = "0.4.0"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.42"
jupyterlab = "^4.2.5"
graphviz = "^0.20.3"

[tool.robotpy]

# Version of robotpy this project depends on
robotpy_version = "2024.3.2.2"

# Which extra RobotPy components should be installed
# -> equivalent to `pip install robotpy[extra1, ...]
robotpy_extras = [
    # "all",
    "apriltag",
    "commands2",
    # "cscore",
    # "navx",
    # "pathplannerlib",
    # "phoenix5",
    # "phoenix6",
    # "photonvision",
    # "playingwithfusion",
    # "rev",
    # "romi",
    # "sim",
    "xrp",
]

# Other pip packages to install
requires = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
